feat: isolate appearance settings
