+++ src/routes/(logged_in)/settings/+layout.svelte
@@ -9,9 +9,16 @@
   <div id="settings-div">
     <div id="settings-nav">
       <a href="/settings" class:current={$page.url.pathname == '/settings'}>Account</a>
-      <a href="/settings/profile">Profile</a>
-      <a href="/settings/notifications">Notifications</a>
-      <a href="/settings/Appearance">Appearance</a>
+      <a href="/settings/profile" class:current={$page.url.pathname == '/settings/profile'}
+        >Profile</a
+      >
+      <a
+        href="/settings/notifications "
+        class:current={$page.url.pathname == '/settings.notifications'}>Notifications</a
+      >
+      <a href="/settings/appearance" class:current={$page.url.pathname == '/settings/appearance'}
+        >Appearance</a
+      >
       <a id="back-link" href="/">Back</a>
     </div>
     <slot />
@@ -30,8 +37,11 @@
 
 <style>
   #settings-wrapper-div {
+    display: flex;
+    flex-direction: column;
     margin-top: 30px;
     overflow-y: scroll;
+    min-height: calc(100% - 90px);
   }
 
   #settings-div {
@@ -41,6 +51,7 @@
     flex-direction: column;
     align-items: center;
     gap: 20px;
+    flex-grow: 1;
   }
 
   #settings-nav {
@@ -54,6 +65,7 @@
   #settings-nav a {
     border: unset;
     padding: 10px;
+    transition: unset;
   }
 
   #back-link {
@@ -72,6 +84,34 @@
     background-color: var(--gray-200);
     width: 100%;
     box-sizing: border-box;
+  }
+
+  :global(.setting) {
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    width: 100%;
+  }
+
+  :global(.setting > label) {
+    font-size: 20px;
+    display: inline-block;
+  }
+
+  :global(.setting input),
+  :global(.setting textarea) {
+    font-size: 18px;
+    padding: 5px 10px;
+    outline: none;
+    border: 2px solid var(--pink-200);
+    border-radius: 10px;
+    background-color: var(--purple-200);
+    color: inherit;
+    resize: none;
+  }
+
+  :global(.setting button) {
+    width: fit-content;
   }
 
   @media only screen and (max-width: 1200px) {
